restructured auth component
