+++ client/src/Assets/MagnifyIcon.png
