+++ client/src/Features/Auth/components/Auth0ProviderWithHistory.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { useNavigate } from 'react-router-dom';
 import { Auth0Provider } from '@auth0/auth0-react';
 
-const Auth0ProviderWithHistory = ({ children }) => {
+export const Auth0ProviderWithHistory = ({ children }) => {
   const domain = process.env.REACT_APP_AUTH0_DOMAIN;
   const clientId = process.env.REACT_APP_AUTH0_CLIENT_ID;
 
@@ -12,7 +12,6 @@ const Auth0ProviderWithHistory = ({ children }) => {
     navigate(appState?.returnTo || window.location.pathname);
   };
 
-  console.log('domain',domain)
   return (
     <Auth0Provider
       domain={domain}
@@ -24,5 +23,3 @@ const Auth0ProviderWithHistory = ({ children }) => {
     </Auth0Provider>
   );
 };
-
-export default Auth0ProviderWithHistory;
