+++ client/src/Features/Auth/components/AuthenticationButton.js
@@ -5,10 +5,8 @@ import {LogoutButton} from './LogoutButton';
 
 import { useAuth0 } from '@auth0/auth0-react';
 
-const AuthenticationButton = () => {
+export const AuthenticationButton = () => {
   const { isAuthenticated } = useAuth0();
 
   return isAuthenticated ? <LogoutButton /> : <LoginButton />;
 };
-
-export default AuthenticationButton;
