+++ client/src/Features/Auth/components/protected-route.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { withAuthenticationRequired } from '@auth0/auth0-react';
 import  {Loading}  from './Loading';
 
-export default function ProtectedRoute ({ component }) {
+export function ProtectedRoute ({ component }) {
     const Component = withAuthenticationRequired(component, {
         onRedirecting: () => (
           <div className="page-layout">
