+++ client/src/Features/Auth/index.js
@@ -1,0 +1,4 @@
+
+export {Auth0ProviderWithHistory} from './components/Auth0ProviderWithHistory'
+export {AuthenticationButton} from './components/AuthenticationButton'
+export {ProtectedRoute} from './components/protected-route'
