+++ client/src/Navigation/Navbar.js
@@ -1,5 +1,5 @@
 import { Link } from "react-router-dom"
-import AuthenticationButton from "../Auth/AuthenticationButton"
+import {AuthenticationButton} from "../Features/Auth"
 import { useAuth0 } from "@auth0/auth0-react";
 import {useState} from 'react'
 import NavModal from './NavModal'
