+++ client/src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import Auth0ProviderWithHistory from './Auth/Auth0ProviderWithHistory';
+import {Auth0ProviderWithHistory} from './Features/Auth/';
 import './index.css';
 import App from './App';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
